added extra error handling

+++ src/system/base/actions/actions.js
@@ -1,7 +1,6 @@
 
 let actions = { };
 
-
 const getActionsFromDb = db => new Promise((resolve, reject) => {
   db.open().then(database => {
     database.all('SELECT * FROM actions', (err, actions) => {
@@ -18,7 +17,6 @@ const saveActionToDb = (db, topic, value) => new Promise((resolve, reject) => {
   db.open().then(database => {
     const query = `INSERT OR REPLACE INTO actions (topic, value) values ('${topic}', '${value}')`;
     database.all(query, (err) => {
-      //database.close();
       if (err){
         reject(err);
       }else{
@@ -45,9 +43,12 @@ const onActionData = (db, topic, value) => new Promise((resolve, reject) => {
 });
 
 const unregisterAction = (db, topic) => new Promise((resolve, reject) => {
+  if (typeof(topic) !== typeof('')){
+    throw (new Error('baseSystem:actions:unregisterAction topic must be a string'));
+  }
+
   db.open().then(database => {
     database.all(`DELETE FROM actions WHERE topic = ('${topic}')`, (err) => {
-      //database.close();
       delete actions[topic];
       if (err){
         reject(err);
